# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Kong Chen",
      "url" : "http://kongch.com",
      "email" : "kongchen@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "users"
  }, {
    "name" : "topics"
  } ],
  "paths" : {
    "/topics" : {
      "get" : {
        "tags" : [ "topics" ],
        "summary" : "Returns specified topics.",
        "description" : "",
        "operationId" : "getTopics",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "searchText",
          "in" : "query",
          "description" : "Text you want to search.",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortBy",
          "in" : "query",
          "description" : "Specifies whether you want to sort by name or date.",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortOrder",
          "in" : "query",
          "description" : "Specifies ascending or descending order for sort.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "topics" ],
        "summary" : "Creates new topic.",
        "description" : "",
        "operationId" : "createTopic",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TopicVO"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/topics/{name}" : {
      "delete" : {
        "tags" : [ "topics" ],
        "summary" : "Deletes topic.",
        "description" : "",
        "operationId" : "deleteTopic",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the topic you want to delete.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/topics/{name}/comments" : {
      "get" : {
        "tags" : [ "topics" ],
        "summary" : "Returns all comment for specified topic.",
        "description" : "",
        "operationId" : "getAllComments",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "topics" ],
        "summary" : "Adds new comment to specified topic.",
        "description" : "",
        "operationId" : "addCommentToTopic",
        "parameters" : [ {
          "name" : "param-creditcard",
          "in" : "header",
          "description" : "Name of the topic you want to add coment to.",
          "required" : true,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the topic you want to add coment to.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Comment you want to add.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CommentVO"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/users" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Returns all users.",
        "description" : "",
        "operationId" : "getAllUsers",
        "produces" : [ "application/xml", "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Creates new user.",
        "description" : "",
        "operationId" : "createUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserVO"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/users/{username}" : {
      "delete" : {
        "tags" : [ "users" ],
        "summary" : "Deletes specified user.",
        "description" : "",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "CommentVO" : {
      "type" : "object",
      "properties" : {
        "authorUsername" : {
          "type" : "string",
          "description" : "Username of the comment's author."
        },
        "text" : {
          "type" : "string",
          "description" : "Text of the comment."
        }
      },
      "description" : "Comment object."
    },
    "TopicVO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "text" : {
          "type" : "string"
        }
      }
    },
    "UserVO" : {
      "type" : "object",
      "properties" : {
        "username" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "surname" : {
          "type" : "string"
        }
      }
    }
  }
}